+++ src/templates/index.js
@@ -1,5 +1,5 @@
 export * from './menu.js';
-export * from './subway_station.js';
-export * from './subway_line.js';
-export * from './subway_section.js';
-export * from './subway_map.js';
+export * from './station.js';
+export * from './line.js';
+export * from './section.js';
+export * from './map.js';
